fix: Int values from the CLI
This PR partially (for ints/floats) fixes #2722.

+++ tests/validate_test.py
@@ -69,7 +69,7 @@ def test_set_virt_file_size():
     result = validate.validate_virt_file_size("8")
 
     # Assert
-    assert isinstance(result, int)
+    assert isinstance(result, float)
     assert result == 8
